+++ src/config.php
@@ -11,5 +11,7 @@ return [
     'requireShippingAddressAtCheckout' => false,
     'requireBillingAddressAtCheckout' => false,
     'pdfPaperSize' => 'letter',
-    'pdfPaperOrientation' => 'portrait'
+    'pdfPaperOrientation' => 'portrait',
+    'autoSetLastUsedShippingAddressOnNewCarts' => true,
+    'autoSetLastUsedBillingAddressOnNewCarts' => true
 ];